DatabaseConnection refactor
