+++ src/FileSystem/StateManagement/FileSystemSnapshotter.cs
@@ -25,14 +25,14 @@ namespace DevOptimal.SystemUtilities.FileSystem.StateManagement
         public ISnapshot SnapshotDirectory(string path)
         {
             var originator = new DirectoryOriginator(path, fileSystem);
-            var snapshot = new Caretaker<DirectoryOriginator, DirectoryMemento>(originator, this);
+            var snapshot = new Caretaker<DirectoryOriginator, DirectoryMemento>(originator, connection);
             return snapshot;
         }
 
         public ISnapshot SnapshotFile(string path)
         {
             var originator = new FileOriginator(path, fileCache, fileSystem);
-            var snapshot = new Caretaker<FileOriginator, FileMemento>(originator, this);
+            var snapshot = new Caretaker<FileOriginator, FileMemento>(originator, connection);
             return snapshot;
         }
     }
