+++ test/Environment.Tests/StateManagement/EnvironmentSnapshotterTests.cs
@@ -1,6 +1,5 @@
 ﻿using DevOptimal.SystemUtilities.Environment.StateManagement;
 using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Threading.Tasks;
